refactor(app): extract runtime media events
